requirements.txt
markdown2==2.4.10
jinja2==3.1.2
python-frontmatter==1.0.0
PyYAML==6.0.1
click==8.1.7
